Provisioning: Split Migrate stat-card components into separate files
Pure structural refactor of the Migrate to GitOps cards slice. Moves the
inline stats logic and card components out of Migrate.tsx into their own
files. No behavior change.
