+++ public/app/features/provisioning/Migrate/StatCard.tsx
@@ -1,0 +1,137 @@
+import { css, cx } from '@emotion/css';
+
+import { type GrafanaTheme2 } from '@grafana/data';
+import { Icon, type IconName, Text, useStyles2 } from '@grafana/ui';
+
+export type StatTone = 'neutral' | 'success' | 'info' | 'warning' | 'primary';
+
+interface StatCardProps {
+  icon: IconName;
+  tone: StatTone;
+  big: string;
+  subLabel?: string;
+  label: string;
+  emphasized?: boolean;
+}
+
+export function StatCard({ icon, tone, big, subLabel, label, emphasized }: StatCardProps) {
+  const styles = useStyles2(getStyles);
+  return (
+    <div
+      className={cx(
+        styles.statCard,
+        styles[`statCardSurface_${tone}` as const],
+        emphasized && styles.statCardEmphasized
+      )}
+    >
+      <div className={cx(styles.statCardIcon, styles[`statIconTone_${tone}` as const])}>
+        <Icon name={icon} size="xl" />
+      </div>
+      <div className={styles.statCardBody}>
+        <span className={cx(styles.statCardLabel, styles[`statCardTone_${tone}` as const])}>{label}</span>
+        <span className={styles.statCardValue}>{big}</span>
+        {subLabel && (
+          <Text color="secondary" variant="body">
+            {subLabel}
+          </Text>
+        )}
+      </div>
+    </div>
+  );
+}
+
+const getStyles = (theme: GrafanaTheme2) => ({
+  statCard: css({
+    display: 'flex',
+    flexDirection: 'row',
+    gap: theme.spacing(2),
+    padding: theme.spacing(2),
+    borderRadius: theme.shape.radius.default,
+    border: `1px solid ${theme.colors.border.medium}`,
+    background: theme.colors.background.secondary,
+    alignItems: 'center',
+    boxShadow: theme.shadows.z1,
+  }),
+  statCardSurface_neutral: css({}),
+  statCardSurface_success: css({
+    background: theme.colors.success.transparent,
+    borderColor: theme.colors.success.borderTransparent,
+  }),
+  statCardSurface_info: css({
+    background: theme.colors.info.transparent,
+    borderColor: theme.colors.info.borderTransparent,
+  }),
+  statCardSurface_warning: css({
+    background: theme.colors.warning.transparent,
+    borderColor: theme.colors.warning.borderTransparent,
+  }),
+  statCardSurface_primary: css({
+    background: `color-mix(in srgb, ${theme.visualization.getColorByName('purple')} 12%, ${theme.colors.background.secondary})`,
+    borderColor: `color-mix(in srgb, ${theme.visualization.getColorByName('purple')} 35%, transparent)`,
+  }),
+  statCardEmphasized: css({
+    borderColor: theme.colors.warning.border,
+    boxShadow: `0 0 0 1px ${theme.colors.warning.borderTransparent}, ${theme.shadows.z1}`,
+  }),
+  statCardBody: css({
+    display: 'flex',
+    flexDirection: 'column',
+    gap: theme.spacing(0.25),
+    minWidth: 0,
+  }),
+  statCardIcon: css({
+    flex: '0 0 auto',
+    width: theme.spacing(6),
+    height: theme.spacing(6),
+    borderRadius: theme.shape.radius.circle,
+    display: 'inline-flex',
+    alignItems: 'center',
+    justifyContent: 'center',
+  }),
+  statIconTone_neutral: css({
+    background: theme.colors.background.canvas,
+    color: theme.colors.text.secondary,
+  }),
+  statIconTone_success: css({
+    background: theme.colors.success.transparent,
+    color: theme.colors.success.text,
+  }),
+  statIconTone_info: css({
+    background: theme.colors.info.transparent,
+    color: theme.colors.info.text,
+  }),
+  statIconTone_warning: css({
+    background: theme.colors.warning.transparent,
+    color: theme.colors.warning.text,
+  }),
+  statIconTone_primary: css({
+    background: `color-mix(in srgb, ${theme.visualization.getColorByName('purple')} 20%, transparent)`,
+    color: theme.visualization.getColorByName('purple'),
+  }),
+  statCardLabel: css({
+    fontSize: theme.typography.body.fontSize,
+    fontWeight: theme.typography.fontWeightMedium,
+  }),
+  statCardValue: css({
+    fontSize: '44px',
+    lineHeight: 1.1,
+    fontWeight: theme.typography.fontWeightBold,
+    color: theme.colors.text.primary,
+    letterSpacing: '-0.5px',
+  }),
+  statCardTone_neutral: css({
+    color: theme.colors.text.primary,
+  }),
+  statCardTone_success: css({
+    color: theme.colors.success.text,
+  }),
+  statCardTone_info: css({
+    color: theme.colors.info.text,
+  }),
+  statCardTone_warning: css({
+    color: theme.colors.warning.text,
+  }),
+  statCardTone_primary: css({
+    color: theme.visualization.getColorByName('purple'),
+  }),
+});
